chore(@mlc/lib): move api contracts into dir;
